+++ common/src/main/java/software/bluelib/loader/animation/AnimationController.java
@@ -19,7 +19,7 @@ import org.jetbrains.annotations.Nullable;
 import software.bluelib.api.utils.logging.BaseLogLevel;
 import software.bluelib.api.utils.logging.BaseLogger;
 import software.bluelib.loader.animatable.base.BlueAnimatable;
-import software.bluelib.loader.animation.bone.BoneSnapshot;
+import software.bluelib.loader.animation.keyframe.BoneFrame;
 import software.bluelib.loader.animation.keyframe.InterpolationData;
 import software.bluelib.loader.animation.keyframe.BoneAnimationFrame;
 import software.bluelib.loader.animation.keyframe.KeyframeLocation;
@@ -53,7 +53,7 @@ public class AnimationController<T extends BlueAnimatable> {
 	@NotNull
 	protected final Map<String, BoneAnimationFrame> boneAnimationQueues = new Object2ObjectOpenHashMap<>();
 	@NotNull
-	protected final Map<String, BoneSnapshot> boneSnapshots = new Object2ObjectOpenHashMap<>();
+	protected final Map<String, BoneFrame> boneSnapshots = new Object2ObjectOpenHashMap<>();
 	@NotNull
 	protected Queue<AnimationProcessor.QueuedAnimation> animationQueue = new LinkedList<>();
 
@@ -308,7 +308,7 @@ public class AnimationController<T extends BlueAnimatable> {
 		return this.stateHandler.handle(pState);
 	}
 
-	public void process(@NotNull BlueModel<T> pModel, @NotNull AnimationState<T> pState, @NotNull Map<String, BoneCache> pBones, @NotNull Map<String, BoneSnapshot> pSnapshots, final double pSeekTime, boolean pCrashWhenCantFindBone) {
+	public void process(@NotNull BlueModel<T> pModel, @NotNull AnimationState<T> pState, @NotNull Map<String, BoneCache> pBones, @NotNull Map<String, BoneFrame> pSnapshots, final double pSeekTime, boolean pCrashWhenCantFindBone) {
 		double adjustedTick = adjustTick(pSeekTime);
 		this.lastModel = pModel;
 
@@ -367,10 +367,10 @@ public class AnimationController<T extends BlueAnimatable> {
 
 				for (BoneAnimationCache boneAnimationCache : this.currentAnimation.animationCache().boneAnimationCaches()) {
 					BoneAnimationFrame boneAnimationFrame = this.boneAnimationQueues.get(boneAnimationCache.boneName());
-					BoneSnapshot boneSnapshot = this.boneSnapshots.get(boneAnimationCache.boneName());
+					BoneFrame boneFrame = this.boneSnapshots.get(boneAnimationCache.boneName());
 					BoneCache bone = pBones.get(boneAnimationCache.boneName());
 
-					if (boneSnapshot == null)
+					if (boneFrame == null)
 						continue;
 
 					if (bone == null) {
@@ -385,21 +385,21 @@ public class AnimationController<T extends BlueAnimatable> {
 					KeyframeStackCache<KeyframeCache<MathValue>> scaleKeyFrames = boneAnimationCache.scaleKeyFrames();
 
 					if (!rotationKeyFrames.xKeyframes().isEmpty()) {
-						boneAnimationFrame.addNextRotation(null, adjustedTick, this.transitionLength, boneSnapshot, bone.getInitialSnapshot(),
+						boneAnimationFrame.addNextRotation(null, adjustedTick, this.transitionLength, boneFrame, bone.getInitialSnapshot(),
 								getAnimationPointAtTick(rotationKeyFrames.xKeyframes(), 0, true, Axis.X),
 								getAnimationPointAtTick(rotationKeyFrames.yKeyframes(), 0, true, Axis.Y),
 								getAnimationPointAtTick(rotationKeyFrames.zKeyframes(), 0, true, Axis.Z));
 					}
 
 					if (!positionKeyFrames.xKeyframes().isEmpty()) {
-						boneAnimationFrame.addNextPosition(null, adjustedTick, this.transitionLength, boneSnapshot,
+						boneAnimationFrame.addNextPosition(null, adjustedTick, this.transitionLength, boneFrame,
 								getAnimationPointAtTick(positionKeyFrames.xKeyframes(), 0, false, Axis.X),
 								getAnimationPointAtTick(positionKeyFrames.yKeyframes(), 0, false, Axis.Y),
 								getAnimationPointAtTick(positionKeyFrames.zKeyframes(), 0, false, Axis.Z));
 					}
 
 					if (!scaleKeyFrames.xKeyframes().isEmpty()) {
-						boneAnimationFrame.addNextScale(null, adjustedTick, this.transitionLength, boneSnapshot,
+						boneAnimationFrame.addNextScale(null, adjustedTick, this.transitionLength, boneFrame,
 								getAnimationPointAtTick(scaleKeyFrames.xKeyframes(), 0, false, Axis.X),
 								getAnimationPointAtTick(scaleKeyFrames.yKeyframes(), 0, false, Axis.Y),
 								getAnimationPointAtTick(scaleKeyFrames.zKeyframes(), 0, false, Axis.Z));
@@ -528,12 +528,12 @@ public class AnimationController<T extends BlueAnimatable> {
 		}
 	}
 
-	private void saveSnapshotsForAnimation(@NotNull AnimationProcessor.QueuedAnimation pAnimation, @NotNull Map<String, BoneSnapshot> pSnapshots) {
-		for (BoneSnapshot snapshot : pSnapshots.values()) {
+	private void saveSnapshotsForAnimation(@NotNull AnimationProcessor.QueuedAnimation pAnimation, @NotNull Map<String, BoneFrame> pSnapshots) {
+		for (BoneFrame snapshot : pSnapshots.values()) {
 			if (pAnimation.animationCache().boneAnimationCaches() != null) {
 				for (BoneAnimationCache boneAnimationCache : pAnimation.animationCache().boneAnimationCaches()) {
 					if (boneAnimationCache.boneName().equals(snapshot.getBone().getName())) {
-						this.boneSnapshots.put(boneAnimationCache.boneName(), BoneSnapshot.copy(snapshot));
+						this.boneSnapshots.put(boneAnimationCache.boneName(), BoneFrame.copy(snapshot));
 
 						break;
 					}