+++ common/src/main/java/software/bluelib/loader/animation/AnimationProcessor.java
@@ -21,7 +21,7 @@ import software.bluelib.api.utils.logging.BaseLogLevel;
 import software.bluelib.api.utils.logging.BaseLogger;
 import software.bluelib.loader.animatable.base.AnimatableManager;
 import software.bluelib.loader.animatable.base.BlueAnimatable;
-import software.bluelib.loader.animation.bone.BoneSnapshot;
+import software.bluelib.loader.animation.keyframe.BoneFrame;
 import software.bluelib.loader.animation.keyframe.InterpolationData;
 import software.bluelib.loader.animation.keyframe.BoneAnimationFrame;
 import software.bluelib.loader.animation.math.Easing;
@@ -72,7 +72,7 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 	}
 
 	public void tickAnimation(@NotNull T pAnimatable, @NotNull BlueModel<T> pModel, @NotNull AnimatableManager<T> pAnimatableManager, double pAnimTime, @NotNull AnimationState<T> pState, boolean pCrashWhenCantFindBone) {
-		Map<String, BoneSnapshot> boneSnapshots = updateBoneSnapshots(pAnimatableManager.getBoneSnapshotCollection());
+		Map<String, BoneFrame> boneSnapshots = updateBoneSnapshots(pAnimatableManager.getBoneSnapshotCollection());
 
 		for (AnimationController<T> controller : pAnimatableManager.getAnimationControllers().values()) {
 			if (this.reloadAnimations) {
@@ -87,19 +87,19 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 			controller.process(pModel, pState, this.bones, boneSnapshots, pAnimTime, pCrashWhenCantFindBone);
 
 			for (BoneAnimationFrame boneAnimation : controller.getBoneAnimationQueues().values()) {
-				BoneCache bone = boneAnimation.bone();
-				BoneSnapshot snapshot = boneSnapshots.get(bone.getName());
-				BoneSnapshot initialSnapshot = bone.getInitialSnapshot();
+				BoneCache bone = boneAnimation.getBone();
+				BoneFrame snapshot = boneSnapshots.get(bone.getName());
+				BoneFrame initialSnapshot = bone.getInitialSnapshot();
 
-				InterpolationData rotXPoint = boneAnimation.rotation().x().poll();
-				InterpolationData rotYPoint = boneAnimation.rotation().y().poll();
-				InterpolationData rotZPoint = boneAnimation.rotation().z().poll();
-				InterpolationData posXPoint = boneAnimation.position().x().poll();
-				InterpolationData posYPoint = boneAnimation.position().y().poll();
-				InterpolationData posZPoint = boneAnimation.position().z().poll();
-				InterpolationData scaleXPoint = boneAnimation.scale().x().poll();
-				InterpolationData scaleYPoint = boneAnimation.scale().y().poll();
-				InterpolationData scaleZPoint = boneAnimation.scale().z().poll();
+				InterpolationData rotXPoint = boneAnimation.getRotation().x().poll();
+				InterpolationData rotYPoint = boneAnimation.getRotation().y().poll();
+				InterpolationData rotZPoint = boneAnimation.getRotation().z().poll();
+				InterpolationData posXPoint = boneAnimation.getPosition().x().poll();
+				InterpolationData posYPoint = boneAnimation.getPosition().y().poll();
+				InterpolationData posZPoint = boneAnimation.getPosition().z().poll();
+				InterpolationData scaleXPoint = boneAnimation.getScale().x().poll();
+				InterpolationData scaleYPoint = boneAnimation.getScale().y().poll();
+				InterpolationData scaleZPoint = boneAnimation.getScale().z().poll();
 				Easing easing = controller.overrideEasingTypeFunction.apply(pAnimatable);
 
 				if (rotXPoint != null && rotYPoint != null && rotZPoint != null) {
@@ -136,8 +136,8 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 
 		for (BoneCache bone : getRegisteredBones()) {
 			if (!bone.hasRotationChanged()) {
-				BoneSnapshot initialSnapshot = bone.getInitialSnapshot();
-				BoneSnapshot saveSnapshot = boneSnapshots.get(bone.getName());
+				BoneFrame initialSnapshot = bone.getInitialSnapshot();
+				BoneFrame saveSnapshot = boneSnapshots.get(bone.getName());
 
 				if (saveSnapshot.isRotAnimInProgress())
 					saveSnapshot.stopRotAnim(pAnimTime);
@@ -178,8 +178,8 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 			}
 
 			if (!bone.hasPositionChanged()) {
-				BoneSnapshot initialSnapshot = bone.getInitialSnapshot();
-				BoneSnapshot saveSnapshot = boneSnapshots.get(bone.getName());
+				BoneFrame initialSnapshot = bone.getInitialSnapshot();
+				BoneFrame saveSnapshot = boneSnapshots.get(bone.getName());
 
 				if (saveSnapshot.isPosAnimInProgress())
 					saveSnapshot.stopPosAnim(pAnimTime);
@@ -195,8 +195,8 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 			}
 
 			if (!bone.hasScaleChanged()) {
-				BoneSnapshot initialSnapshot = bone.getInitialSnapshot();
-				BoneSnapshot saveSnapshot = boneSnapshots.get(bone.getName());
+				BoneFrame initialSnapshot = bone.getInitialSnapshot();
+				BoneFrame saveSnapshot = boneSnapshots.get(bone.getName());
 
 				if (saveSnapshot.isScaleAnimInProgress())
 					saveSnapshot.stopScaleAnim(pAnimTime);
@@ -228,10 +228,10 @@ public class AnimationProcessor<T extends BlueAnimatable> {
 	}
 
 	@NotNull
-	private Map<String, BoneSnapshot> updateBoneSnapshots(@NotNull Map<String, BoneSnapshot> pSnapshots) {
+	private Map<String, BoneFrame> updateBoneSnapshots(@NotNull Map<String, BoneFrame> pSnapshots) {
 		for (BoneCache bone : getRegisteredBones()) {
 			if (!pSnapshots.containsKey(bone.getName()))
-				pSnapshots.put(bone.getName(), BoneSnapshot.copy(bone.getInitialSnapshot()));
+				pSnapshots.put(bone.getName(), BoneFrame.copy(bone.getInitialSnapshot()));
 		}
 
 		return pSnapshots;