+++ common/src/main/java/software/bluelib/loader/animation/keyframe/BoneAnimationFrame.java
@@ -2,22 +2,26 @@ package software.bluelib.loader.animation.keyframe;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import software.bluelib.loader.animation.bone.BoneSnapshot;
 import software.bluelib.loader.animation.keyframe.frame.AnimationFrameVector;
 import software.bluelib.loader.cache.animations.keyframe.KeyframeCache;
 import software.bluelib.loader.cache.model.BoneCache;
 
-public record BoneAnimationFrame(@NotNull BoneCache bone,
-                                 @NotNull AnimationFrameVector rotation,
-                                 @NotNull AnimationFrameVector position,
-                                 @NotNull AnimationFrameVector scale) {
+public final class BoneAnimationFrame {
+
+	private final BoneCache bone;
+	private final AnimationFrameVector rotation;
+	private final AnimationFrameVector position;
+	private final AnimationFrameVector scale;
 
 	public BoneAnimationFrame(@NotNull BoneCache pBone) {
-		this(pBone, new AnimationFrameVector(), new AnimationFrameVector(), new AnimationFrameVector());
+		this.bone = pBone;
+		this.rotation = new AnimationFrameVector();
+		this.position = new AnimationFrameVector();
+		this.scale = new AnimationFrameVector();
 	}
 
 	public void addNextPosition(@Nullable KeyframeCache<?> pKeyFrame, double pLerpedTick, double pTransitionLength,
-	                            @NotNull BoneSnapshot pStartSnapshot,
+	                            @NotNull BoneFrame pStartSnapshot,
 	                            @NotNull InterpolationData pNextX, @NotNull InterpolationData pNextY, @NotNull InterpolationData pNextZ) {
 		position.addPoint(pKeyFrame, pLerpedTick, pTransitionLength,
 				pStartSnapshot.getOffsetX(), pNextX.startValue(),
@@ -26,7 +30,7 @@ public record BoneAnimationFrame(@NotNull BoneCache bone,
 	}
 
 	public void addNextScale(@Nullable KeyframeCache<?> pKeyFrame, double pLerpedTick, double pTransitionLength,
-	                         @NotNull BoneSnapshot pStartSnapshot,
+	                         @NotNull BoneFrame pStartSnapshot,
 	                         @NotNull InterpolationData pNextX, @NotNull InterpolationData pNextY, @NotNull InterpolationData pNextZ) {
 		scale.addPoint(pKeyFrame, pLerpedTick, pTransitionLength,
 				pStartSnapshot.getScaleX(), pNextX.startValue(),
@@ -35,7 +39,7 @@ public record BoneAnimationFrame(@NotNull BoneCache bone,
 	}
 
 	public void addNextRotation(@Nullable KeyframeCache<?> pKeyFrame, double pLerpedTick, double pTransitionLength,
-	                            @NotNull BoneSnapshot pStartSnapshot, @NotNull BoneSnapshot pInitialSnapshot,
+	                            @NotNull BoneFrame pStartSnapshot, @NotNull BoneFrame pInitialSnapshot,
 	                            @NotNull InterpolationData pNextX, @NotNull InterpolationData pNextY, @NotNull InterpolationData pNextZ) {
 		rotation.addPoint(pKeyFrame, pLerpedTick, pTransitionLength,
 				pStartSnapshot.getRotX() - pInitialSnapshot.getRotX(), pNextX.startValue(),
@@ -53,5 +57,21 @@ public record BoneAnimationFrame(@NotNull BoneCache bone,
 
 	public void addNextScale(@NotNull InterpolationData pX, @NotNull InterpolationData pY, @NotNull InterpolationData pZ) {
 		scale.add(pX, pY, pZ);
+	}
+
+	public BoneCache getBone() {
+		return bone;
+	}
+
+	public AnimationFrameVector getRotation() {
+		return rotation;
+	}
+
+	public AnimationFrameVector getPosition() {
+		return position;
+	}
+
+	public AnimationFrameVector getScale() {
+		return scale;
 	}
 }